+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## next
+
+- Fixed code highlighting in `source` view
+
 ## 1.0.0-beta.61 (31-03-2021)
 
 - Fixed preloader's block disability in some cases due to underlaying