+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added alias `jora` for `discovery-query` syntax (used in `source` and `markdown`)
 - Added anchor support for headers and `anchor` option to enable it
 - Added anchor support for headers in `markdown` view and `anchors` option to disable it
+- Changed `hstack` to use flexbox for layout
 
 ## 1.0.0-beta.61 (31-03-2021)