+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## next
+## 1.0.0-beta.62 (14-10-2021)
 
 - Fixed code highlighting in `source` view
 - Added alias `jora` for `discovery-query` syntax (used in `source` and `markdown`)